bot/requirements.txt
aiogram
PyGithub
python-dotenv
qrcode
Pillow
aiohttp-socks
